+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/PointerModel.cs
@@ -198,6 +198,7 @@ namespace UnityEngine.InputSystem.UI
         {
             private bool m_IsPressed;
             private PointerEventData.FramePressState m_FramePressState;
+            private float m_PressTime;
 
             public bool isPressed
             {
@@ -218,27 +219,40 @@ namespace UnityEngine.InputSystem.UI
                 }
             }
 
+            public float pressTime
+            {
+                get => m_PressTime;
+                set => m_PressTime = value;
+            }
+
+            public bool clickedOnSameGameObject
+            {
+                get => m_ClickedOnSameGameObject;
+                set => m_ClickedOnSameGameObject = value;
+            }
+
             public bool wasPressedThisFrame => m_FramePressState == PointerEventData.FramePressState.Pressed ||
             m_FramePressState == PointerEventData.FramePressState.PressedAndReleased;
             public bool wasReleasedThisFrame => m_FramePressState == PointerEventData.FramePressState.Released ||
             m_FramePressState == PointerEventData.FramePressState.PressedAndReleased;
 
-            private RaycastResult pressRaycast;
-            private GameObject pressObject;
-            private GameObject rawPressObject;
-            private GameObject lastPressObject;
-            private GameObject dragObject;
-            private Vector2 pressPosition;
-            private float clickTime; // On Time.unscaledTime timeline, NOT input event time.
-            private int clickCount;
-            private bool dragging;
+            private RaycastResult m_PressRaycast;
+            private GameObject m_PressObject;
+            private GameObject m_RawPressObject;
+            private GameObject m_LastPressObject;
+            private GameObject m_DragObject;
+            private Vector2 m_PressPosition;
+            private float m_ClickTime; // On Time.unscaledTime timeline, NOT input event time.
+            private int m_ClickCount;
+            private bool m_Dragging;
+            private bool m_ClickedOnSameGameObject;
 
             public void CopyPressStateTo(PointerEventData eventData)
             {
-                eventData.pointerPressRaycast = pressRaycast;
-                eventData.pressPosition = pressPosition;
-                eventData.clickCount = clickCount;
-                eventData.clickTime = clickTime;
+                eventData.pointerPressRaycast = m_PressRaycast;
+                eventData.pressPosition = m_PressPosition;
+                eventData.clickCount = m_ClickCount;
+                eventData.clickTime = m_ClickTime;
                 // We can't set lastPress directly. Old input module uses three different event instances, one for each
                 // button. We share one instance and just switch press states. Set pointerPress twice to get the lastPress
                 // we need.
@@ -246,24 +260,24 @@ namespace UnityEngine.InputSystem.UI
                 // NOTE: This does *NOT* quite work as stated in the docs. pointerPress is nulled out on button release which
                 //       will set lastPress as a side-effect. This means that lastPress will only be non-null while no press is
                 //       going on and will *NOT* refer to the last pressed object when a new object has been pressed on.
-                eventData.pointerPress = lastPressObject;
-                eventData.pointerPress = pressObject;
-                eventData.rawPointerPress = rawPressObject;
-                eventData.pointerDrag = dragObject;
-                eventData.dragging = dragging;
+                eventData.pointerPress = m_LastPressObject;
+                eventData.pointerPress = m_PressObject;
+                eventData.rawPointerPress = m_RawPressObject;
+                eventData.pointerDrag = m_DragObject;
+                eventData.dragging = m_Dragging;
             }
 
             public void CopyPressStateFrom(PointerEventData eventData)
             {
-                pressRaycast = eventData.pointerPressRaycast;
-                pressObject = eventData.pointerPress;
-                rawPressObject = eventData.rawPointerPress;
-                lastPressObject = eventData.lastPress;
-                pressPosition = eventData.pressPosition;
-                clickTime = eventData.clickTime;
-                clickCount = eventData.clickCount;
-                dragObject = eventData.pointerDrag;
-                dragging = eventData.dragging;
+                m_PressRaycast = eventData.pointerPressRaycast;
+                m_PressObject = eventData.pointerPress;
+                m_RawPressObject = eventData.rawPointerPress;
+                m_LastPressObject = eventData.lastPress;
+                m_PressPosition = eventData.pressPosition;
+                m_ClickTime = eventData.clickTime;
+                m_ClickCount = eventData.clickCount;
+                m_DragObject = eventData.pointerDrag;
+                m_Dragging = eventData.dragging;
             }
 
             public void OnEndFrame()